requirements.txt
torch
torchvision
torchaudio
transformers
datasets
sentencepiece
accelerate
scikit-learn
pillow
pandas
numpy
